refactor: extract checkSyntax plugin as a standalone package
