+++ .changeset/new-crews-marry.md
@@ -1,0 +1,6 @@
+---
+'@rsbuild/shared': patch
+'@rsbuild/core': patch
+---
+
+refactor: extract checkSyntax plugin as a standalone package
