+++ e2e/cases/source/source-include/index.test.ts
@@ -1,16 +1,15 @@
 import path from 'path';
 import { expect, test } from '@playwright/test';
 import { build } from '@scripts/shared';
+import { pluginCheckSyntax } from '@rsbuild/plugin-check-syntax';
 
 test('should not compile file which outside of project by default', async () => {
   await expect(
     build({
       cwd: __dirname,
       entry: { index: path.resolve(__dirname, './src/index.js') },
+      plugins: [pluginCheckSyntax()],
       rsbuildConfig: {
-        security: {
-          checkSyntax: true,
-        },
         output: {
           overrideBrowserslist: ['ie 11'],
         },
@@ -24,12 +23,10 @@ test('should compile specified file when source.include', async () => {
     build({
       cwd: __dirname,
       entry: { index: path.resolve(__dirname, './src/index.js') },
+      plugins: [pluginCheckSyntax()],
       rsbuildConfig: {
         source: {
           include: [path.resolve(__dirname, '../test.js')],
-        },
-        security: {
-          checkSyntax: true,
         },
       },
     }),
