+++ packages/compat/webpack/src/shared/plugin.ts
@@ -34,7 +34,6 @@ export const applyDefaultPlugins = (plugins: Plugins) =>
     import('../plugins/hmr').then((m) => m.pluginHMR()),
     plugins.svg(),
     import('../plugins/pug').then((m) => m.pluginPug()),
-    plugins.checkSyntax(),
     import('../plugins/copy').then((m) => m.pluginCopy()),
     import('../plugins/react').then((m) => m.pluginReactWebpack()),
     plugins.font(),
