+++ packages/plugin-check-syntax/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @rsbuild/plugin-check-syntax
