+++ packages/plugin-check-syntax/src/helpers/getEcmaVersion.ts
