+++ packages/plugin-check-syntax/src/helpers/index.ts
@@ -1,4 +1,3 @@
-export * from './type';
 export * from './generateError';
 export * from './printErrors';
 export * from './getEcmaVersion';
