+++ packages/shared/package.json
@@ -163,22 +163,18 @@
     "@modern-js/server": "^2.39.0",
     "@types/babel__core": "^7.20.3",
     "@types/fs-extra": "^11.0.2",
-    "acorn": "^8.10.0",
     "browserslist": "^4.22.1",
-    "caniuse-lite": "^1.0.30001520",
     "chalk": "^4.1.2",
     "css-minimizer-webpack-plugin": "5.0.1",
     "cssnano": "6.0.1",
     "deepmerge": "^4.3.1",
     "fs-extra": "^11.1.1",
-    "htmlparser2": "9.0.0",
     "line-diff": "2.1.1",
     "lodash": "^4.17.21",
     "pkg-up": "^3.1.0",
     "postcss": "8.4.31",
     "rslog": "^1.1.0",
     "semver": "^7.5.4",
-    "source-map": "^0.7.4",
     "url-join": "^4.0.1",
     "upath": "2.0.1",
     "webpack": "^5.88.1",
@@ -191,7 +187,6 @@
     "@babel/core": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
-    "@types/caniuse-lite": "^1.0.1",
     "@types/lodash": "^4.14.200",
     "@types/node": "^16",
     "@types/semver": "^7.5.4",
