--- packages/shared/src/plugins/CheckSyntaxPlugin/helpers/type.ts
@@ -1,37 +1,0 @@
-import { CheckSyntaxOptions } from '../../../types';
-
-export type CheckSyntaxExclude = NonNullable<CheckSyntaxOptions['exclude']>;
-
-export interface Location {
-  line: number;
-  column: number;
-}
-
-export interface File {
-  path: string;
-  line: number;
-  column: number;
-}
-
-interface SyntaxErrorOptions {
-  source: File & { code: string };
-  output?: File;
-}
-
-export class SyntaxError extends Error {
-  source: File & { code: string };
-
-  output: File | undefined;
-
-  constructor(message: string, options: SyntaxErrorOptions) {
-    super(message);
-    this.source = options.source;
-    this.output = options.output;
-  }
-}
-
-export type AcornParseError = {
-  message: string;
-  pos: number;
-  loc: Location;
-};
